refactor: Moved methods used in multiple places to a separate file
