+++ app/src/main/java/com/android/solvit/shared/ui/authentication/ForgotPassword.kt
@@ -44,6 +44,8 @@ import androidx.compose.ui.unit.sp
 import coil.compose.rememberAsyncImagePainter
 import com.android.solvit.R
 import com.android.solvit.shared.ui.navigation.NavigationActions
+import com.android.solvit.shared.ui.utils.CustomOutlinedTextField
+import com.android.solvit.shared.ui.utils.ValidationRegex
 
 /**
  * A composable function that provides a "Forgot Password" screen for users to reset their password.
